Move testing and add "any dataframe" detector
