+++ sklearn/utils/_array_api.py
@@ -16,7 +16,7 @@ from sklearn._config import get_config
 from sklearn.externals import array_api_compat
 from sklearn.externals import array_api_extra as xpx
 from sklearn.externals.array_api_compat import numpy as np_compat
-from sklearn.utils._dataframe import is_pandas_df_or_series
+from sklearn.utils._dataframe import is_df_or_series
 from sklearn.utils.fixes import parse_version
 
 # TODO: complete __all__
@@ -321,7 +321,7 @@ def _remove_non_arrays(*arrays, remove_none=True, remove_types=(str,)):
             continue
         if sp.issparse(array):
             continue
-        if is_pandas_df_or_series(array):
+        if is_df_or_series(array):
             continue
         filtered_arrays.append(array)
